fix: refactor to have less duplication
